# Improve performance of ILI9341 on ATSAMD5X