I've made performance improvements.

1. clear+set in write8() is now set only.
    * Instead of being faster, it's more restrictive than ever before.
    * In the original, it had to be an 8-bit contiguous port, but this time it has to be an 8-bit contiguous port accessible by StoreUint8().
    * If this restriction is unacceptable, it is better not to merge these PRs.