@sago35 2020-04-17
test code : 7e992534be5ba34b5cba9cfba5ca0e28

* FillRectangle()
    * before : 35ms (100%)
    * with 1 : 28ms (80%)
    * with 1 and 2 : 23ms (66%)

* DrawRGBBitmap()
    * before : 83ms (100%)
    * with 1 : 59ms (71%)
    * with 1 and 2 : 52ms (63%)