@sago35 2020-04-17
This PR will certainly make you faster, but it will also increase your limitations.
If it looks like it would be better to merge, I'll consider organizing the code a bit more.

Do you have any thoughts on that?