@bgould 2020-04-18
I haven't tried this extensively but this change seems to reduce the fps of the pyportal_boing example slightly... not sure if it is specific to that program, or if LLVM is doing some magic like unrolling the loop itself more efficiently, or something else.  Do you have that hardware? Are you able to see a similar result?