@sago35 2020-04-18
pyportal_boing I tried it.
I wasn't aware of it, but it certainly slows down.

* pyportal_boing demo
    * before : 42 fps
    * with 1 : 51 fps
    * with 1 and 3 : 42 fps

There may be some kind of loop optimization being done in LLVM.
Interestingly, when the first loop is set to +=4 instead of +=8, it is initially 51 fps, but gradually drops to around 44 fps.
I don't know why.

I'll check a bit more if I know how to see the final assembler code.