@aykevl 2020-04-26
LLVM does a lot of optimizations (including loop unrolling and loop re-rolling) that may all interact with each other. I've sometimes managed to speed things up or slown things down by changes that should have had the opposite effect. In general, making the code simpler to understand usually also makes it easier to understand for the compiler and thus easier to optimize.

In my experience, one of the few things that reliably increases performance is avoiding I/O operations. Those have side effects unknown to the compiler so cannot be optimized by it. See for example https://github.com/tinygo-org/drivers/pull/89. Another that may help is to avoid branches (if/else), those are not always well optimized by LLVM.

Also, if you run performance sensitive code, remember that TinyGo allows tuning the optimization level. By default it uses `-opt=z` meaning optimize for code size at all cost. You can try `-opt=s` (optimize for size without impacting performance too much) and `-opt=2` (optimize for performance). This setting can have quite a significant effect on the performance of code, but also on code size.