@sago35 2020-04-27
I understood.

I think it's better to leave 2 and 3 to LLVM optimization.
1 is exactly what reduces the I/O operations.

I think this PR could be completed by making it just 1.
I force-pushed the PR.


benchmarks: 
tinygo build -o app.uf2 -target pyportal -size short -opt=* ./examples/ili9341/pyportal_boing

* before (-opt=z) : 42 fps
* before (-opt=2) : 38 - 42 fps
* before (-opt=s) : 38 - 42 fps
* with1 (-opt=z) : 51 fps
* with 1 (-opt=2) : 47 - 51 fps
* with 1 (-opt=s) : 47 - 51 fps