@bgould 2020-04-28
Yep I tried this out with the latest push and there is a noticeable improvement.  Nice work, that is a good speedup IMO.  I'm in favor of this change