@deadprogram 2020-05-06
Finally got a chance to test this, and it looks great! :globe_with_meridians: 

Thank you @sago35 now merging.